// apps/sage-sync/src/lib/sage/types.ts
type SageRecord = {
  RECORD_MODIFY_DATE: string;
  RECORD_CREATE_DATE: string;
  RECORD_DELETED: 0 | 1;
};
// ...
export type SALES_LEDGER = SageRecord & {
  ACCOUNT_REF: string;
  NAME: string;
};

export type SALES_ORDER = SageRecord & {
  ORDER_NUMBER: number;
  ACCOUNT_REF: string;
  ORDER_DATE: string;
  ORDER_OR_QUOTE: string;
  ORDER_TYPE_CODE: number;
  DESPATCH_DATE: string | null;
  DESPATCH_STATUS: string;
  DESPATCH_STATUS_CODE: number;
};

export type SOP_ITEM = SageRecord & {
  ITEMID: number;
  ORDER_NUMBER: number;
  STOCK_CODE: string;
  DELIVERY_DATE: string;
  DUE_DATE: string | null;
  QTY_ALLOCATED: number;
  QTY_DELIVERED: number;
  QTY_DESPATCH: number;
  QTY_ORDER: number;
};